+++ src/edu/tdp2/client/dto/MyVendedorAccount.java
@@ -11,16 +11,16 @@ public class MyVendedorAccount extends MySpecificAccount implements Serializable
 {
 	private static final long serialVersionUID = -6866869107363384714L;
 
-	private Map<Moneda, Long> gananciaAcumulada;
+	private Map<Moneda, Double> gananciaAcumulada;
 	private List<Proyecto> proyectosPerdidos;
 	private List<Proyecto> proyectosAdjudicados;
 
-	public Map<Moneda, Long> getGananciaAcumulada()
+	public Map<Moneda, Double> getGananciaAcumulada()
 	{
 		return gananciaAcumulada;
 	}
 
-	public void setGananciaAcumulada(Map<Moneda, Long> gananciaAcumulada)
+	public void setGananciaAcumulada(Map<Moneda, Double> gananciaAcumulada)
 	{
 		this.gananciaAcumulada = gananciaAcumulada;
 	}
