+++ src/edu/tdp2/client/model/Oferta.java
@@ -21,7 +21,7 @@ public class Oferta extends AbstractDomainObject
 	private static final long serialVersionUID = -4981789655579048475L;
 
 	@Column(name = "Monto", nullable = false)
-	private int monto;
+	private float monto;
 
 	@Column(name = "CantDias", nullable = false)
 	private int dias;
@@ -93,12 +93,12 @@ public class Oferta extends AbstractDomainObject
 		this.proyecto = proyecto;
 	}
 
-	public int getMonto()
+	public float getMonto()
 	{
 		return monto;
 	}
 
-	public void setMonto(int monto)
+	public void setMonto(float monto)
 	{
 		this.monto = monto;
 	}
