+++ src/edu/tdp2/client/widgets/DetailSearchWidget.java
@@ -173,7 +173,7 @@ public class DetailSearchWidget extends VerticalPanel {
 		};
 		ClientUtils.getSoftmartService().isUsuarioBloqueado(oferta.getUsuario().getLogin(), callback);
 
-		table.setWidget(row, 0, new HTML(Integer.toString(oferta.getMonto())));
+		table.setWidget(row, 0, new HTML(Float.toString(oferta.getMonto())));
 		table.setWidget(row, 1, new HTML(oferta.getMoneda().getDescription()));
 		table.setWidget(row, 2, new HTML(Integer.toString(oferta.getDias())));
